+++ packages/spacecat-shared-data-access/test/unit/service/index.test.js
@@ -100,7 +100,7 @@ describe('Data Access Object Tests', () => {
   ];
 
   const apiKeyFunctions = [
-    'getApiKeyByHashedKey',
+    'getApiKeyByHashedApiKey',
     'createNewApiKey',
   ];