+++ packages/spacecat-shared-http-utils/src/auth/handlers/scoped-api-key.js
@@ -43,11 +43,13 @@ export default class ScopedApiKeyHandler extends AbstractHandler {
       this.log(`No API key entity found in the data layer for the provided API key: ${apiKeyFromHeader}`, 'error');
       return null;
     }
+    this.log(`Valid API key entity found. Id: ${apiKeyEntity.getId()}, name: ${apiKeyEntity.getName()}, scopes: ${apiKeyEntity.getScopes()}`, 'debug');
 
     // We have an API key entity, and need to check if it's still valid
     const authInfo = new AuthInfo()
       .withProfile(apiKeyEntity) // Include the API key entity as the profile
       .withType(this.name);
+    this.log('Successfully constructed authInfo object', 'debug');
 
     // Verify that the api key has not expired or been revoked
     const now = new Date().toISOString();